Remove ConfigKind type
This patch removes the ConfigKind::Frozen mechansim, which wasn't
exported through the public interface at all.

Because the ::Frozen variant was removed, the ConfigKind::Mutable
variant is the only one remaining and because the ConfigKind type isn't
exported in the API as well, we can move the variant members to the
Config struct itself.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>