Reimplement the Config building mechanism
This patch rewrites the Config building mechanism using special objects
for tracking the config building state.

Transitions between states are done on the fly as required.

This is required so that the async sources can be stored inside the
configuration building objects, while keeping out the expenses in the
non-async case, so a user of the crate has only to pay for what they are
using (no async means no overhead for that).

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
Reviewed-by: Matthias Beyer <mail@beyermatthias.de>