errors: clarify names of integer types
I got this error message:

```
invalid type: integer 64 bit `5`, expected an array
```

It took me a while to figure out that it wasn't talking about bit
number 5 of the number 64, but the 64-bit number `5`. I think it will
be clearer like this:

```
invalid type: 64-bit integer `5`, expected an array
```

So that's what this patch implements.