Remove "watch" example
This patch removes the "watch" example. The idea should be straight
forward to a decently experienced rust dev, and (to be honest) I am too
lazy right now to update the example for the removed (because
deprecated) interfaces.

Shame on me.

If someone wants to provide this example again, I'm happy to see this
patch reverted and the example adapted. ;-)

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>