Run clippy only on MSRV
This patch reverts the matrix for the clippy job to only run clippy on
our MSRV instead of stable, beta and nightly.

The recurring issue with this was that lints started failing in PRs that
couldn't have been fixed in master yet, because master was never tested
with a clippy that new (in case of nightly).

1.56.1 is good enough for now.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
(cherry picked from commit 336a34e52c8b0889180c865b29c89420cd1788dc)