Fix test for error string
This patch fixes the error test for parsing invalid TOML.
When we updated toml-rs to 0.7.0, the error display string was altered.
Thus we had to adapt the tests for that string.

The legacy tests were removed as they are equal to the "non-legacy"
tests. No need to keep them both.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>