Use Cargo.lock.msrv in CI tests
This will prevent our CI breaking due to updates to our dependencies.

Because of the way the msrv.yml file is constructed, it was not
straightforward to use the committed lockfile only for the MSRV tests.

It would be better to use a different lockfile, with latest versions
of our dependencies, for non-MSRV tests.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>